+++ packages/ui/src/demo-mocks/network/mutations/promote-jobs.ts
@@ -1,0 +1,7 @@
+import type { PromoteJobsMutation } from '@/typings/gql';
+
+export const promoteJobsMock = (): Promise<PromoteJobsMutation> => {
+  return Promise.resolve({
+    promoteJobs: [],
+  });
+};